fix: 解决 src/api 路径下，不能导出非 http methods 函数
